requirements.txt
torch
torchvision
Pillow
numpy
scikit-learn
transformers
flask
joblib
flask_cors
